vdesk-agents/pyproject.toml
[project]
name = "vdesk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.1.0",
    "paho-mqtt>=2.1.0",
    "python-dotenv",
]
